refactor: unnecessary hook
