+++ packages/shared/src/hooks/utils/index.ts
@@ -1,3 +1,2 @@
 export * from './usePagination';
 export * from './useThemedAsset';
-export * from './useDownloadUrl';
